+++ src/main/java/vn/edu/fpt/petworldplatform/repository/PetRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import vn.edu.fpt.petworldplatform.entity.Pets;
 import java.util.List;
 
-public interface PetRepo extends JpaRepository<Pets, Long> {
+public interface PetRepo extends JpaRepository<Pets, Integer> {
 
     List<Pets> findByNameContaining(String name);
     List<Pets> findByOwner_CustomerId(Long customerId);
